refactor : purchaseAmount 자료형 long으로 변경

+++ src/main/java/lotto/constant/PurchaseAmount.java
@@ -5,13 +5,13 @@ public enum PurchaseAmount {
     MIN(1000),
     MAX(100000);
 
-    private final int amount;
+    private final long amount;
 
-    PurchaseAmount(int amount) {
+    PurchaseAmount(long amount) {
         this.amount = amount;
     }
 
-    public int getAmount() {
+    public long getAmount() {
         return this.amount;
     }
 }
